moved functions to utils
