+++ supplier-retail/src/pages/api/auth/[...nextauth].js
@@ -3,8 +3,8 @@ import NextAuth from 'next-auth'
 import Providers from 'next-auth/providers'
 import Adapters from 'next-auth/adapters'
 import prisma from '../../../../lib/prisma'
-import getUserByEmail from './getUserByRole' 
-import login from './login' 
+import getUserByEmail from '../../../../utils/getUserByRole' 
+import login from '../../../../utils/login' 
 // import { PrismaClient } from '@prisma/client'
 
 // const prisma = new PrismaClient()
