+++ supplier-retail/utils/getUserByRole.ts
@@ -1,4 +1,4 @@
-import prisma from '../../../../lib/prisma'
+import prisma from '../lib/prisma'
 export default async function getUserByEmail( email ) {
 let user = await prisma.user.findUnique({
     where: {
